refactor: format currencies values
