+++ lib/src/presenter/screens/wallet/widgets/assets_list_view.dart
@@ -2,8 +2,10 @@ import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:projeto_criptos/src/presenter/riverpod/user_provider.dart';
 import 'package:projeto_criptos/src/presenter/riverpod/visible_provider.dart';
+import 'package:projeto_criptos/utils/currency_formater.dart';
 
 import '../../../../domain/entities/asset_model.dart';
+import 'crypto_list_tile.dart';
 
 class AssetsListView extends HookConsumerWidget {
   const AssetsListView({
@@ -21,64 +23,16 @@ class AssetsListView extends HookConsumerWidget {
         itemCount: user.state.assets.length,
         itemBuilder: (context, index) {
           AssetModel asset = user.state.assets[index];
-          double cryptoBalance = asset.coinBalance * asset.price;
+          String balance = formatter.format(asset.coinBalance * asset.price);
           bool variation = asset.variation > 0;
           return Column(
             children: [
               const Divider(thickness: 1),
-              ListTile(
-                leading: CircleAvatar(
-                  radius: 20,
-                  backgroundColor: Colors.grey.shade800,
-                  child: Icon(
-                    asset.icon,
-                    color: Colors.white,
-                  ),
-                ),
-                title: Text(asset.symbol),
-                subtitle: Text(
-                  asset.name,
-                  style: TextStyle(
-                    fontWeight: FontWeight.w500,
-                    color: Colors.grey.shade600,
-                  ),
-                ),
-                trailing: Column(
-                  crossAxisAlignment: CrossAxisAlignment.end,
-                  mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                  children: [
-                    Text(
-                      'R\$ ${visible.state ? cryptoBalance.toStringAsFixed(2) : '*******'}',
-                      style: const TextStyle(
-                        fontWeight: FontWeight.w400,
-                        fontSize: 15,
-                      ),
-                    ),
-                    Container(
-                      padding: const EdgeInsets.symmetric(
-                        horizontal: 10,
-                        vertical: 3,
-                      ),
-                      decoration: BoxDecoration(
-                        borderRadius: BorderRadius.circular(15),
-                        color: variation
-                            ? const Color.fromARGB(255, 134, 249, 230)
-                            : Colors.red.shade100,
-                      ),
-                      child: Text(
-                        '${variation ? '+' : ''} ${asset.variation}%',
-                        style: TextStyle(
-                          color: variation
-                              ? const Color.fromARGB(255, 47, 129, 116)
-                              : Colors.red,
-                          fontSize: 12,
-                          fontWeight: FontWeight.w500,
-                        ),
-                        textAlign: TextAlign.center,
-                      ),
-                    ),
-                  ],
-                ),
+              CryptoListTile(
+                asset: asset,
+                visible: visible,
+                balance: balance,
+                variation: variation,
               ),
             ],
           );
