+++ lib/utils/currency_formater.dart
@@ -1,0 +1,3 @@
+import 'package:intl/intl.dart';
+
+NumberFormat formatter = NumberFormat("#,##0.00", "pt-BR");
